# internal/transport/http/dto/activity.go
package dto

import "time"

// ActivityDTO represents the activity as exposed to the transport layer.
type ActivityDTO struct {
	ID          uint   `json:"id"`
	Name        string `json:"name"`
	Description string `json:"description"`
	CreatedAt   time.Time `json:"created_at"`
	UpdatedAt   time.Time `json:"updated_at"`
}